feat: delete coordinator and user

+++ src/useCases/coordinator/deleteCoordinatorUseCase.ts
@@ -1,3 +1,4 @@
+import { AppError } from '@/errors';
 import { prisma } from '@/lib/prisma';
 
 interface DeleteCoordinatorUseCaseRequest {
@@ -6,8 +7,23 @@ interface DeleteCoordinatorUseCaseRequest {
 
 export class DeleteCoordinatorUseCase {
   async execute({ id }: DeleteCoordinatorUseCaseRequest) {
+    const coordinator = await prisma.coordinator.findUnique({
+      where: { id },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!coordinator) {
+      throw new AppError('Usuário inexistente', 400);
+    }
+
     await prisma.coordinator.delete({
       where: { id },
+    });
+
+    await prisma.user.delete({
+      where: { id: coordinator.userId },
     });
   }
 }
